+++ src/shared/interfaces/ApiResponse.ts
@@ -1,5 +1,8 @@
 export type ApiResponse<Key extends string, T> = {
   [K in Key]: {
     results: T[];
+    info: {
+      next: number | null;
+    };
   };
 };